api/requirements.txt
asgiref==3.7.2
Django==4.1.10
pip==23.2.1
setuptools==60.2.0
typing-extensions==4.7.1
tzdata==2023.3
wheel==0.37.1
Scrapy~=2.9.0
djongo==1.3.6
pytz==2023.3
pymongo==3.12.3
google-cloud-secret-manager==2.16.2